# gob-simulator/src/main/resources/openapi/libra-gateway-hearing-events-v0.3.0.yml
...
openapi: "3.0.3"
info:
  title: "Libra Gateway Hearing Event API"
  version: "0.3.0"
  description: "REST API exposed by Libra Gateway for Common Platform hearing confirmation and hearing result callbacks. Hearing event paths do not carry an external path identifier; the request payload supplies the case/account context."
servers:
  -
    url: "https://cpp-gob.gateway.lncs.hmcs"
    description: "Production Libra Gateway endpoint."
  -
    url: "https://libra-cpp-gob-gateway.staging.internal.hmcts.net"
    description: "Pre-Production Libra Gateway endpoint."
...
paths:
...
  "/hearing":
    post:
      tags:
        - "Hearing Events"
      operationId: "confirmHearing"
      summary: "Confirm an enforcement hearing"
      description: "Receives allocated hearing details from Common Platform, resolves the GoB enforcement record from the payload context, removes the enforcement hold and records the hearing details."
      security:
        -
          ClientCredentials:
            []
      parameters:
        -
          $ref: "#/components/parameters/CorrelationIdHeader"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/HearingConfirmedRequest"
            example:
              caseUrn: "E012345678"
              courtHearingLocation: "B01BH01"
              dateOfHearing: "2026-04-24"
              timeOfHearing: "14:00"
      responses:
        "200":
          description: "Hearing confirmation updated successfully."
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  schemas:
...
    HearingConfirmedRequest:
      type: "object"
      additionalProperties: false
      properties:
        caseUrn:
          type: "string"
          maxLength: 36
          description: "This will be the Case URN that was provided by GOB/LG in SubmitCase Business rules: Provided in the format \u2013 \u2018Sequence: XX00\u2026......\u2019 Unique value for each submitCase request. Used as the unifying value in any subsequent messages and communications. Must be 10 characters, an E with 9 numbers E.g. E012345678 CP comment: Example: \"ENFA654789B\" Source: HearingConfirmed row 2 (from extracted workbook JSON)"
          x-source-row: 2
          x-source-data-attribute: "caseUrn"
        courtHearingLocation:
          type: "string"
          maxLength: 7
          description: "CJSE Organisational Unit code identifying the destination court (and courtroom where applicable). Mapped from GoB Enforcement Court Code via national court code mapping; manual enforcement appends selected courtroom, auto enforcement uses '00'. Business rules: DS: Court Location : OUCODE (Complete Court location including Courtroom) Will display on the front screen on account in GOB in Ac History CP comment: Format - OUCODE (Example: \"B01BH01\") Source: HearingConfirmed row 3 (from extracted workbook JSON)"
          x-source-row: 3
          x-source-data-attribute: "courtHearingLocation"
        dateOfHearing:
          type: "string"
          format: "date"
          description: "Date of the target Court Hearing DS : YYYY-MM-DD Business rules: This can be 31 days in the past CP comment: format - 'YYYY-MM-DD' (Example: \"2026-04-24\") Source: HearingConfirmed row 4 (from extracted workbook JSON)"
          x-source-row: 4
          x-source-data-attribute: "dateOfHearing"
        timeOfHearing:
          type: "string"
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
          maxLength: 5
          description: "The actual start time of the target Court Hearing HH:MM Business rules: 24 Hour clock i.e. 10:00 and 14:00 This will be the actual scheduled time e.g. if the Hearing date is 12th August 2026 at 14:00 and the available slot is 12th August 15:30 then hearing is confirmed allocation at 15:30. CP comment: 24hrs clock. (Example: \"14:00\") Source: HearingConfirmed row 5 (from extracted workbook JSON)"
          x-source-row: 5
          x-source-data-attribute: "timeOfHearing"
      x-source: "HearingConfirmed worksheet"
      description: "Common Platform callback used to confirm that an enforcement hearing has been allocated."
      required:
        - "caseUrn"
        - "courtHearingLocation"
...
    ErrorResponse:
      type: "object"
      additionalProperties: false
      required:
        - "errorCode"
        - "errorDescription"
      properties:
        errorCode:
          type: "string"
        errorDescription:
          type: "string"
  securitySchemes:
    ClientCredentials:
      type: "oauth2"
      flows:
        clientCredentials:
          tokenUrl: "/auth/token"
          scopes:
            {}
  parameters:
    CorrelationIdHeader:
      name: "X-Correlation-ID"
      in: "header"
      required: false
      description: "Optional caller-supplied correlation identifier for traceability."
      schema:
        type: "string"
  responses:
    BadRequest:
      description: "Validation error, invalid payload, or invalid status transition."
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: "Missing, expired or invalid bearer token."
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: "Authenticated client is not authorised to invoke the operation."
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: "No GoB enforcement record can be resolved for the supplied case/account context."
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: "Unexpected Gateway, DAC or GoB processing error."
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorResponse"